requirements.txt
beautifulsoup4==4.10.0
lxml==4.6.3
matplotlib==3.3.4
nest_asyncio==1.5.1
nltk==3.6.5
numpy==1.21.4
pandas==1.2.4
Pillow==8.4.0
requests==2.26.0
scikit_learn==1.0.1
seaborn==0.11.1
selectorlib==0.16.0
selenium==4.0.0b4
textblob==0.17.1
twint==2.1.20
undetected_chromedriver==3.0.3
webdriver_manager==3.5.2
wordcloud==1.8.1
